HID: input: do not run GET_REPORT unless there's a Resolution Multiplier
Some devices take GET_REPORT as an instruction to change the mode, or
toggle some other value, or possibly do unspeakable things to kittens.
So we must not execute GET_REPORT against a device unless we're sure
that there's a Resolution Multiplier in that report that makes it all
worth it.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>